http

+++ SmartHome/PackManager.m
@@ -13,7 +13,7 @@
 + (NSData *) fireflyProtocol:(NSString *)cmd
 {
     NSData* bytes = [cmd dataUsingEncoding:NSUTF8StringEncoding];
-    int len=[cmd length]+4;
+    long len=[cmd length]+4;
     Byte array[]={0,0,0,0,0,0,0,0,0,0,0,0,len,0,0,0,2};
     NSData *data = [NSData dataWithBytes: array length: sizeof(array)];
     NSMutableData *ret=[[NSMutableData alloc] initWithData:data];
